Move common tests and test utils out of main package
