+++ jest.common.mjs
@@ -1,7 +1,15 @@
+// These options are imported into the Jest config of each package, so
+// `<rootDir>` refers to the respective /packages/* folder, e.g.
+// /packages/antd, /packages/material, etc., not the root of the repo.
+
 /** @type {import('@jest/types').Config.InitialOptions} */
 export default {
   coveragePathIgnorePatterns: ['/genericTests/', '/dist/', 'TestUtils.ts'],
-  // This path is relative to the /packages/* folders
-  setupFilesAfterEnv: ['../../jestSetup.ts'],
+  setupFilesAfterEnv: ['<rootDir>/../../jestSetup.ts'],
   testEnvironment: 'jsdom',
+  // Keep these in sync with /tsconfig.json#compilerOptions#paths.
+  moduleNameMapper: {
+    'react-querybuilder': ['<rootDir>/../../packages/react-querybuilder/src'],
+    '@rqb-test': ['<rootDir>/../../utils/genericTests'],
+  },
 };
