+++ package.json
@@ -38,7 +38,7 @@
     "lint": "eslint 'packages/*/src/**/*.{ts,tsx}' 'examples/**/*.{js,ts,tsx}' 'website/{docs,src,static}/**/*.{js,ts,tsx}'",
     "test": "concurrently --group --timings --names sqlite,jsonlogic,jsonata,postgres,jest \"bun test dbquery.sqlite\" \"bun test dbquery.jsonlogic\" \"bun test dbquery.jsonata\" \"bun test dbquery.postgres\" jest",
     "test:watch": "jest --watch",
-    "typecheck": "concurrently --timings --max-processes 14 --prefix-colors \"#0958d9,#712cf9,#00d1b2,#81e6d9,#4078c0,#7fba00,#339af0,#007fff,#61dafb,#82a7dd,#60a5fa,gray,blue\" --names antd,bootstrap,bulma,chakra,dnd,fluent,mantine,material,native,rqb,tremor,examples,website bun:typecheck:antd bun:typecheck:bootstrap bun:typecheck:bulma bun:typecheck:chakra bun:typecheck:dnd bun:typecheck:fluent bun:typecheck:mantine bun:typecheck:material bun:typecheck:native bun:typecheck:rqb bun:typecheck:tremor bun:typecheck:examples bun:typecheck:website",
+    "typecheck": "concurrently --timings --max-processes 14 --prefix-colors \"#0958d9,#712cf9,#00d1b2,#81e6d9,#4078c0,#7fba00,#339af0,#007fff,#61dafb,#82a7dd,#60a5fa,darkgray,gray,blue\" --names antd,bootstrap,bulma,chakra,dnd,fluent,mantine,material,native,rqb,tremor,utils,examples,website bun:typecheck:antd bun:typecheck:bootstrap bun:typecheck:bulma bun:typecheck:chakra bun:typecheck:dnd bun:typecheck:fluent bun:typecheck:mantine bun:typecheck:material bun:typecheck:native bun:typecheck:rqb bun:typecheck:tremor bun:typecheck:utils bun:typecheck:examples bun:typecheck:website",
     "typecheck:antd": "tsc -p packages/antd",
     "typecheck:bootstrap": "tsc -p packages/bootstrap",
     "typecheck:bulma": "tsc -p packages/bulma",
@@ -52,6 +52,7 @@
     "typecheck:tremor": "tsc -p packages/tremor",
     "typecheck:examples": "tsc -p examples",
     "typecheck:website": "tsc -p website",
+    "typecheck:utils": "tsc -p ./tsconfig.json",
     "are-the-types-wrong": "bun run attw:rqb && bun run attw:antd && bun run attw:bootstrap && bun run attw:bulma && bun run attw:chakra && bun run attw:dnd && bun run attw:fluent && bun run attw:mantine && bun run attw:material && bun run attw:native && bun run attw:tremor",
     "attw:antd": "attw --format table-flipped --pack ./packages/antd",
     "attw:bootstrap": "attw --format table-flipped --pack ./packages/bootstrap",
@@ -64,8 +65,8 @@
     "attw:native": "attw --format table-flipped --pack ./packages/native",
     "attw:rqb": "attw --format table-flipped --exclude-entrypoints 'dev/*' 'dist/*' 'genericTests/*' --ignore-rules false-export-default --pack ./packages/react-querybuilder",
     "attw:tremor": "attw --format table-flipped --pack ./packages/tremor",
-    "pretty-print": "prettier --config prettier.config.mjs --write '*.{mj,t}s' 'packages/*/{src,dev,genericTests}/**' 'website/*.js' 'website/{docs,src,versioned_docs,versioned_sidebars}/**'",
-    "pretty-check": "prettier --config prettier.config.mjs --check '*.{mj,t}s' 'packages/*/{src,dev,genericTests}/**' 'website/*.js' 'website/{docs,src,versioned_docs,versioned_sidebars}/**'",
+    "pretty-print": "prettier --config prettier.config.mjs --write '*.{mj,t}s' './packages/*/{src,dev}/**' './utils/**' './website/*.js' './website/{docs,src,versioned_docs,versioned_sidebars}/**'",
+    "pretty-check": "prettier --config prettier.config.mjs --check '*.{mj,t}s' './packages/*/{src,dev}/**' './utils/**' './website/*.js' './website/{docs,src,versioned_docs,versioned_sidebars}/**'",
     "generate-examples": "bun ./examples/generateExamples.ts",
     "generate-parsers": "bun run --filter react-querybuilder generate:parsers",
     "outdated": "bunx npm-check-updates --interactive --format=group --workspaces --root",
