+++ packages/bulma/src/Bulma.test.tsx
@@ -1,11 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import * as React from 'react';
 import { QueryBuilder, TestID } from 'react-querybuilder';
-import {
-  testNotToggle,
-  testValueEditor,
-  testValueSelector,
-} from 'react-querybuilder/genericTests/index';
+import { testNotToggle, testValueEditor, testValueSelector } from '@rqb-test';
 import { BulmaNotToggle } from './BulmaNotToggle';
 import { BulmaValueEditor } from './BulmaValueEditor';
 import { BulmaValueSelector } from './BulmaValueSelector';
