+++ packages/dnd/src/RuleDnD.test.tsx
@@ -16,7 +16,7 @@ import type {
   RuleGroupType,
 } from 'react-querybuilder';
 import { QueryBuilder, TestID, standardClassnames as sc } from 'react-querybuilder';
-import { userEventSetup } from 'react-querybuilder/genericTests/index';
+import { userEventSetup } from '@rqb-test';
 import { QueryBuilderDnD } from './QueryBuilderDnD';
 import type { QueryBuilderDndProps } from './types';
