+++ packages/react-querybuilder/src/components/ActionElement.test.ts
@@ -1,4 +1,4 @@
-import { testActionElement } from '../../genericTests';
+import { testActionElement } from '@rqb-test';
 import { ActionElement } from './ActionElement';
 
 testActionElement(ActionElement);
