+++ packages/react-querybuilder/src/components/DragHandle.test.ts
@@ -1,4 +1,4 @@
-import { testDragHandle } from '../../genericTests';
+import { testDragHandle } from '@rqb-test';
 import { DragHandle } from './DragHandle';
 
 testDragHandle(DragHandle);
