+++ packages/react-querybuilder/src/components/NotToggle.test.ts
@@ -1,4 +1,4 @@
-import { testNotToggle } from '../../genericTests';
+import { testNotToggle } from '@rqb-test';
 import { NotToggle } from './NotToggle';
 
 testNotToggle(NotToggle);
