+++ packages/react-querybuilder/src/components/Rule.test.tsx
@@ -7,7 +7,7 @@ import {
   getFieldMapFromArray,
   getRuleProps as getProps,
   ruleClassnames,
-} from '../../genericTests';
+} from '@rqb-test';
 import { TestID, standardClassnames as sc, defaultTranslations as t } from '../defaults';
 import { messages } from '../messages';
 import type {
