+++ packages/react-querybuilder/src/components/RuleGroup.test.tsx
@@ -1,12 +1,7 @@
+import { consoleMocks, createRule, getRuleGroupProps, ruleGroupClassnames } from '@rqb-test';
 import { screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import * as React from 'react';
-import {
-  consoleMocks,
-  createRule,
-  getRuleGroupProps,
-  ruleGroupClassnames,
-} from '../../genericTests';
 import {
   TestID,
   defaultCombinators,
