+++ packages/react-querybuilder/src/components/ShiftActions.test.ts
@@ -1,4 +1,4 @@
-import { testShiftActions } from '../../genericTests';
+import { testShiftActions } from '@rqb-test';
 import { ShiftActions } from './ShiftActions';
 
 testShiftActions(ShiftActions);
