+++ packages/react-querybuilder/src/components/ValueEditor.test.ts
@@ -1,4 +1,4 @@
-import { testValueEditor } from '../../genericTests';
+import { testValueEditor } from '@rqb-test';
 import { ValueEditor } from './ValueEditor';
 
 testValueEditor(ValueEditor);
