+++ packages/react-querybuilder/src/components/ValueSelector.test.ts
@@ -1,4 +1,4 @@
-import { testValueSelector } from '../../genericTests';
+import { testValueSelector } from '@rqb-test';
 import { ValueSelector } from './ValueSelector';
 
 testValueSelector(ValueSelector);
