+++ packages/react-querybuilder/src/utils/prepareQueryObjects.test.ts
@@ -1,5 +1,5 @@
-import { uuidV4regex } from '../../genericTests/generateIDtests';
 import type { RuleGroupType, RuleGroupTypeIC } from '../types/index.noReact';
+import { uuidV4regex } from './generateIDTestUtils';
 import { prepareRule, prepareRuleGroup, prepareRuleOrGroup } from './prepareQueryObjects';
 
 describe('prepareRule', () => {
