+++ packages/react-querybuilder/src/utils/regenerateIDs.test.ts
@@ -1,5 +1,5 @@
-import { uuidV4regex } from '../../genericTests/generateIDtests';
 import type { RuleGroupType, RuleGroupTypeIC, RuleType } from '../types/index.noReact';
+import { uuidV4regex } from './generateIDTestUtils';
 import { regenerateID, regenerateIDs } from './regenerateIDs';
 
 const ruleGroup: RuleGroupType = {
