+++ packages/tremor/src/Tremor.test.tsx
@@ -19,7 +19,7 @@ import {
   testShiftActions,
   testValueEditor,
   userEventSetup,
-} from 'react-querybuilder/genericTests/index';
+} from '@rqb-test';
 import { TremorActionElement } from './TremorActionElement';
 import { TremorNotToggle } from './TremorNotToggle';
 import { TremorShiftActions } from './TremorShiftActions';
