--- utils/dev/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "dev",
-  "version": "0.0.0",
-  "private": true,
-  "exports": {
-    ".": {
-      "import": "./index.ts",
-      "types": "./index.ts"
-    }
-  }
-}
