+++ utils/genericTests/testShiftActions.tsx
@@ -1,7 +1,7 @@
 import { act, render, screen } from '@testing-library/react';
 import * as React from 'react';
-import { TestID } from '../src';
-import type { RuleType, ShiftActionsProps } from '../src/types';
+import type { RuleType, ShiftActionsProps } from 'react-querybuilder';
+import { TestID } from 'react-querybuilder';
 import { basicSchema, userEventSetup } from './utils';
 
 const labels = { shiftUp: 'shiftUp', shiftDown: 'shiftDown' } as const;
